+++ org/src/main/java/cn/gezhi/crm/org/dao/DepartmentMapper.java
@@ -1,0 +1,30 @@
+package cn.gezhi.crm.org.dao;
+
+import cn.gezhi.crm.org.entity.Department;
+import cn.gezhi.crm.org.entity.DepartmentExample;
+import java.util.List;
+import org.apache.ibatis.annotations.Param;
+
+public interface DepartmentMapper {
+    int countByExample(DepartmentExample example);
+
+    int deleteByExample(DepartmentExample example);
+
+    int deleteByPrimaryKey(Integer id);
+
+    int insert(Department record);
+
+    int insertSelective(Department record);
+
+    List<Department> selectByExample(DepartmentExample example);
+
+    Department selectByPrimaryKey(Integer id);
+
+    int updateByExampleSelective(@Param("record") Department record, @Param("example") DepartmentExample example);
+
+    int updateByExample(@Param("record") Department record, @Param("example") DepartmentExample example);
+
+    int updateByPrimaryKeySelective(Department record);
+
+    int updateByPrimaryKey(Department record);
+}
