--- org/src/main/java/cn/gezhi/crm/org/dao/EmployeeMapper.java
@@ -1,33 +1,0 @@
-package cn.gezhi.crm.org.dao;
-
-import cn.gezhi.crm.org.entity.Employee;
-import cn.gezhi.crm.org.entity.EmployeeExample;
-
-import java.util.List;
-
-import org.apache.ibatis.annotations.Param;
-
-public interface EmployeeMapper {
-    int countByExample(EmployeeExample example);
-
-    int deleteByExample(EmployeeExample example);
-
-    int deleteByPrimaryKey(Integer id);
-
-    int insert(Employee record);
-
-    int insertSelective(Employee record);
-
-    List<Employee> selectByExample(EmployeeExample example);
-
-    Employee selectByPrimaryKey(Integer id);
-
-    int updateByExampleSelective(@Param("record") Employee record, @Param("example") EmployeeExample example);
-
-    int updateByExample(@Param("record") Employee record, @Param("example") EmployeeExample example);
-
-    int updateByPrimaryKeySelective(Employee record);
-
-    int updateByPrimaryKey(Employee record);
-
-}
