--- org/src/main/java/cn/gezhi/crm/org/entity/Career.java
@@ -1,45 +1,0 @@
-package cn.gezhi.crm.org.entity;
-
-/**
- * TODO
- *
- * @author pppppap
- */
-public class Career {
-    private Integer id;
-    private String name;
-    private Integer deleteFlag;
-
-    public Integer getId() {
-        return id;
-    }
-
-    public void setId(Integer id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public Integer getDeleteFlag() {
-        return deleteFlag;
-    }
-
-    public void setDeleteFlag(Integer deleteFlag) {
-        this.deleteFlag = deleteFlag;
-    }
-
-    @Override
-    public String toString() {
-        return "Career{" +
-                "id=" + id +
-                ", name='" + name + '\'' +
-                ", deleteFlag=" + deleteFlag +
-                '}';
-    }
-}
