+++ org/src/main/java/cn/gezhi/crm/org/entity/DepartmentExample.java
@@ -1,0 +1,400 @@
+package cn.gezhi.crm.org.entity;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class DepartmentExample {
+    protected String orderByClause;
+
+    protected boolean distinct;
+
+    protected List<Criteria> oredCriteria;
+
+    public DepartmentExample() {
+        oredCriteria = new ArrayList<Criteria>();
+    }
+
+    public void setOrderByClause(String orderByClause) {
+        this.orderByClause = orderByClause;
+    }
+
+    public String getOrderByClause() {
+        return orderByClause;
+    }
+
+    public void setDistinct(boolean distinct) {
+        this.distinct = distinct;
+    }
+
+    public boolean isDistinct() {
+        return distinct;
+    }
+
+    public List<Criteria> getOredCriteria() {
+        return oredCriteria;
+    }
+
+    public void or(Criteria criteria) {
+        oredCriteria.add(criteria);
+    }
+
+    public Criteria or() {
+        Criteria criteria = createCriteriaInternal();
+        oredCriteria.add(criteria);
+        return criteria;
+    }
+
+    public Criteria createCriteria() {
+        Criteria criteria = createCriteriaInternal();
+        if (oredCriteria.size() == 0) {
+            oredCriteria.add(criteria);
+        }
+        return criteria;
+    }
+
+    protected Criteria createCriteriaInternal() {
+        Criteria criteria = new Criteria();
+        return criteria;
+    }
+
+    public void clear() {
+        oredCriteria.clear();
+        orderByClause = null;
+        distinct = false;
+    }
+
+    protected abstract static class GeneratedCriteria {
+        protected List<Criterion> criteria;
+
+        protected GeneratedCriteria() {
+            super();
+            criteria = new ArrayList<Criterion>();
+        }
+
+        public boolean isValid() {
+            return criteria.size() > 0;
+        }
+
+        public List<Criterion> getAllCriteria() {
+            return criteria;
+        }
+
+        public List<Criterion> getCriteria() {
+            return criteria;
+        }
+
+        protected void addCriterion(String condition) {
+            if (condition == null) {
+                throw new RuntimeException("Value for condition cannot be null");
+            }
+            criteria.add(new Criterion(condition));
+        }
+
+        protected void addCriterion(String condition, Object value, String property) {
+            if (value == null) {
+                throw new RuntimeException("Value for " + property + " cannot be null");
+            }
+            criteria.add(new Criterion(condition, value));
+        }
+
+        protected void addCriterion(String condition, Object value1, Object value2, String property) {
+            if (value1 == null || value2 == null) {
+                throw new RuntimeException("Between values for " + property + " cannot be null");
+            }
+            criteria.add(new Criterion(condition, value1, value2));
+        }
+
+        public Criteria andIdIsNull() {
+            addCriterion("id is null");
+            return (Criteria) this;
+        }
+
+        public Criteria andIdIsNotNull() {
+            addCriterion("id is not null");
+            return (Criteria) this;
+        }
+
+        public Criteria andIdEqualTo(Integer value) {
+            addCriterion("id =", value, "id");
+            return (Criteria) this;
+        }
+
+        public Criteria andIdNotEqualTo(Integer value) {
+            addCriterion("id <>", value, "id");
+            return (Criteria) this;
+        }
+
+        public Criteria andIdGreaterThan(Integer value) {
+            addCriterion("id >", value, "id");
+            return (Criteria) this;
+        }
+
+        public Criteria andIdGreaterThanOrEqualTo(Integer value) {
+            addCriterion("id >=", value, "id");
+            return (Criteria) this;
+        }
+
+        public Criteria andIdLessThan(Integer value) {
+            addCriterion("id <", value, "id");
+            return (Criteria) this;
+        }
+
+        public Criteria andIdLessThanOrEqualTo(Integer value) {
+            addCriterion("id <=", value, "id");
+            return (Criteria) this;
+        }
+
+        public Criteria andIdIn(List<Integer> values) {
+            addCriterion("id in", values, "id");
+            return (Criteria) this;
+        }
+
+        public Criteria andIdNotIn(List<Integer> values) {
+            addCriterion("id not in", values, "id");
+            return (Criteria) this;
+        }
+
+        public Criteria andIdBetween(Integer value1, Integer value2) {
+            addCriterion("id between", value1, value2, "id");
+            return (Criteria) this;
+        }
+
+        public Criteria andIdNotBetween(Integer value1, Integer value2) {
+            addCriterion("id not between", value1, value2, "id");
+            return (Criteria) this;
+        }
+
+        public Criteria andDepNameIsNull() {
+            addCriterion("dep_name is null");
+            return (Criteria) this;
+        }
+
+        public Criteria andDepNameIsNotNull() {
+            addCriterion("dep_name is not null");
+            return (Criteria) this;
+        }
+
+        public Criteria andDepNameEqualTo(String value) {
+            addCriterion("dep_name =", value, "depName");
+            return (Criteria) this;
+        }
+
+        public Criteria andDepNameNotEqualTo(String value) {
+            addCriterion("dep_name <>", value, "depName");
+            return (Criteria) this;
+        }
+
+        public Criteria andDepNameGreaterThan(String value) {
+            addCriterion("dep_name >", value, "depName");
+            return (Criteria) this;
+        }
+
+        public Criteria andDepNameGreaterThanOrEqualTo(String value) {
+            addCriterion("dep_name >=", value, "depName");
+            return (Criteria) this;
+        }
+
+        public Criteria andDepNameLessThan(String value) {
+            addCriterion("dep_name <", value, "depName");
+            return (Criteria) this;
+        }
+
+        public Criteria andDepNameLessThanOrEqualTo(String value) {
+            addCriterion("dep_name <=", value, "depName");
+            return (Criteria) this;
+        }
+
+        public Criteria andDepNameLike(String value) {
+            addCriterion("dep_name like", value, "depName");
+            return (Criteria) this;
+        }
+
+        public Criteria andDepNameNotLike(String value) {
+            addCriterion("dep_name not like", value, "depName");
+            return (Criteria) this;
+        }
+
+        public Criteria andDepNameIn(List<String> values) {
+            addCriterion("dep_name in", values, "depName");
+            return (Criteria) this;
+        }
+
+        public Criteria andDepNameNotIn(List<String> values) {
+            addCriterion("dep_name not in", values, "depName");
+            return (Criteria) this;
+        }
+
+        public Criteria andDepNameBetween(String value1, String value2) {
+            addCriterion("dep_name between", value1, value2, "depName");
+            return (Criteria) this;
+        }
+
+        public Criteria andDepNameNotBetween(String value1, String value2) {
+            addCriterion("dep_name not between", value1, value2, "depName");
+            return (Criteria) this;
+        }
+
+        public Criteria andDepDescIsNull() {
+            addCriterion("dep_desc is null");
+            return (Criteria) this;
+        }
+
+        public Criteria andDepDescIsNotNull() {
+            addCriterion("dep_desc is not null");
+            return (Criteria) this;
+        }
+
+        public Criteria andDepDescEqualTo(String value) {
+            addCriterion("dep_desc =", value, "depDesc");
+            return (Criteria) this;
+        }
+
+        public Criteria andDepDescNotEqualTo(String value) {
+            addCriterion("dep_desc <>", value, "depDesc");
+            return (Criteria) this;
+        }
+
+        public Criteria andDepDescGreaterThan(String value) {
+            addCriterion("dep_desc >", value, "depDesc");
+            return (Criteria) this;
+        }
+
+        public Criteria andDepDescGreaterThanOrEqualTo(String value) {
+            addCriterion("dep_desc >=", value, "depDesc");
+            return (Criteria) this;
+        }
+
+        public Criteria andDepDescLessThan(String value) {
+            addCriterion("dep_desc <", value, "depDesc");
+            return (Criteria) this;
+        }
+
+        public Criteria andDepDescLessThanOrEqualTo(String value) {
+            addCriterion("dep_desc <=", value, "depDesc");
+            return (Criteria) this;
+        }
+
+        public Criteria andDepDescLike(String value) {
+            addCriterion("dep_desc like", value, "depDesc");
+            return (Criteria) this;
+        }
+
+        public Criteria andDepDescNotLike(String value) {
+            addCriterion("dep_desc not like", value, "depDesc");
+            return (Criteria) this;
+        }
+
+        public Criteria andDepDescIn(List<String> values) {
+            addCriterion("dep_desc in", values, "depDesc");
+            return (Criteria) this;
+        }
+
+        public Criteria andDepDescNotIn(List<String> values) {
+            addCriterion("dep_desc not in", values, "depDesc");
+            return (Criteria) this;
+        }
+
+        public Criteria andDepDescBetween(String value1, String value2) {
+            addCriterion("dep_desc between", value1, value2, "depDesc");
+            return (Criteria) this;
+        }
+
+        public Criteria andDepDescNotBetween(String value1, String value2) {
+            addCriterion("dep_desc not between", value1, value2, "depDesc");
+            return (Criteria) this;
+        }
+    }
+
+    public static class Criteria extends GeneratedCriteria {
+
+        protected Criteria() {
+            super();
+        }
+    }
+
+    public static class Criterion {
+        private String condition;
+
+        private Object value;
+
+        private Object secondValue;
+
+        private boolean noValue;
+
+        private boolean singleValue;
+
+        private boolean betweenValue;
+
+        private boolean listValue;
+
+        private String typeHandler;
+
+        public String getCondition() {
+            return condition;
+        }
+
+        public Object getValue() {
+            return value;
+        }
+
+        public Object getSecondValue() {
+            return secondValue;
+        }
+
+        public boolean isNoValue() {
+            return noValue;
+        }
+
+        public boolean isSingleValue() {
+            return singleValue;
+        }
+
+        public boolean isBetweenValue() {
+            return betweenValue;
+        }
+
+        public boolean isListValue() {
+            return listValue;
+        }
+
+        public String getTypeHandler() {
+            return typeHandler;
+        }
+
+        protected Criterion(String condition) {
+            super();
+            this.condition = condition;
+            this.typeHandler = null;
+            this.noValue = true;
+        }
+
+        protected Criterion(String condition, Object value, String typeHandler) {
+            super();
+            this.condition = condition;
+            this.value = value;
+            this.typeHandler = typeHandler;
+            if (value instanceof List<?>) {
+                this.listValue = true;
+            } else {
+                this.singleValue = true;
+            }
+        }
+
+        protected Criterion(String condition, Object value) {
+            this(condition, value, null);
+        }
+
+        protected Criterion(String condition, Object value, Object secondValue, String typeHandler) {
+            super();
+            this.condition = condition;
+            this.value = value;
+            this.secondValue = secondValue;
+            this.typeHandler = typeHandler;
+            this.betweenValue = true;
+        }
+
+        protected Criterion(String condition, Object value, Object secondValue) {
+            this(condition, value, secondValue, null);
+        }
+    }
+}
