--- org/src/main/java/cn/gezhi/crm/org/entity/Employee.java
@@ -1,75 +1,0 @@
-package cn.gezhi.crm.org.entity;
-
-public class Employee {
-    private Integer id;
-
-    private Integer careerId;
-
-    private Integer depId;
-
-    private Integer userId;
-
-    private String name;
-
-    private Integer deleteFlag;
-
-    public Integer getId() {
-        return id;
-    }
-
-    public void setId(Integer id) {
-        this.id = id;
-    }
-
-    public Integer getCareerId() {
-        return careerId;
-    }
-
-    public void setCareerId(Integer careerId) {
-        this.careerId = careerId;
-    }
-
-    public Integer getDepId() {
-        return depId;
-    }
-
-    public void setDepId(Integer depId) {
-        this.depId = depId;
-    }
-
-    public Integer getUserId() {
-        return userId;
-    }
-
-    public void setUserId(Integer userId) {
-        this.userId = userId;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name == null ? null : name.trim();
-    }
-
-    public Integer getDeleteFlag() {
-        return deleteFlag;
-    }
-
-    public void setDeleteFlag(Integer deleteFlag) {
-        this.deleteFlag = deleteFlag;
-    }
-
-    @Override
-    public String toString() {
-        return "Employee{" +
-                "id=" + id +
-                ", careerId=" + careerId +
-                ", depId=" + depId +
-                ", userId=" + userId +
-                ", name='" + name + '\'' +
-                ", deleteFlag=" + deleteFlag +
-                '}';
-    }
-}
