+++ org/src/main/java/cn/gezhi/crm/org/service/DepartmentService.java
@@ -1,0 +1,7 @@
+package cn.gezhi.crm.org.service;
+
+import cn.gezhi.crm.org.entity.Department;
+
+public interface DepartmentService {
+ Department getOne(int id);
+}
