+++ org/src/main/java/cn/gezhi/crm/org/service/impl/DepartmentServiceImpl.java
@@ -1,0 +1,28 @@
+package cn.gezhi.crm.org.service.impl;
+
+import cn.gezhi.crm.org.dao.DepartmentMapper;
+import cn.gezhi.crm.org.entity.Department;
+import cn.gezhi.crm.org.service.DepartmentService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+/**
+ * TODO
+ *
+ * @author Jo
+ * @date 2018/8/14
+ */
+@Service
+public class DepartmentServiceImpl implements DepartmentService {
+    @Autowired
+    private DepartmentMapper departmentMapper;
+    /**
+     *TODO 根据ID查询一个部门
+     *@param id 部门的编号
+     *@return  部门对象
+     **/
+
+    public Department getOne(int id) {
+        return departmentMapper.selectByPrimaryKey(id);
+    }
+}
