+++ org/src/main/resources/mybatis-generator-config.xml
@@ -4,7 +4,8 @@
         "http://mybatis.org/dtd/mybatis-generator-config_1_0.dtd" >
 <generatorConfiguration>
     <!-- 本地数据库驱动程序jar包的全路径 -->
-    <classPathEntry location="/home/liu/.m2/repository/mysql/mysql-connector-java/5.1.40/mysql-connector-java-5.1.40.jar"/>
+
+    <classPathEntry location="D:\Maven\Maven_local_Repository\mysql\mysql-connector-java\5.1.40\mysql-connector-java-5.1.40.jar"/>
     <context id="testTables" targetRuntime="MyBatis3">
         <commentGenerator>
             <!-- 是否去除自动生成的注释 true：是 ： false:否 -->
@@ -45,7 +46,7 @@
             <property name="enableSubPackages" value="false" />
         </javaClientGenerator>
         <!-- 指定数据库表 -->
-        <table schema="" tableName="employee"></table>
+        <table schema="" tableName="department"></table>
         <!-- 有些表的字段需要指定java类型
        <table schema="" tableName="">
          <columnOverride column="" javaType="" />
