+++ org/src/test/java/OrgTest.java
@@ -1,5 +1,5 @@
-import cn.gezhi.crm.org.dao.EmployeeMapper;
-import cn.gezhi.crm.org.entity.Employee;
+import cn.gezhi.crm.org.dao.DepartmentMapper;
+import cn.gezhi.crm.org.entity.Department;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,12 +16,27 @@ import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 @ContextConfiguration(locations = {"classpath:spring-placeholder.xml", "classpath:spring/application*.xml"})
 public class OrgTest {
     @Autowired
-    private EmployeeMapper employeeMapper;
-
+    private DepartmentMapper departmentMapper;
+   /**
+    *TODO
+    *@param
+    *@return
+    **/
     @Test
-    public void test1() {
-        Employee employee = employeeMapper.selectByPrimaryKey(1);
-        System.out.println(employee);
+    public void testSelectByPrimaryKey() {
+        Department department = departmentMapper.selectByPrimaryKey(1);
+        System.out.println(department);
+    }
+    @Test
+    public void testInsert() {
+        Department department=new Department();
+        department.setId(5);
+        department.setDepName("外交部");
+        department.setDepDesc("应酬");
+        int insert = departmentMapper.insert(department);
+        System.out.println(insert);
+
+
     }
 
 }
