+++ pom.xml
@@ -15,7 +15,6 @@
     <properties>
         <crm.version>1.0.0</crm.version>
         <spring.version>5.0.7.RELEASE</spring.version>
-        <hibernate.version>5.2.17.Final</hibernate.version>
         <mysql.driver.version>5.1.40</mysql.driver.version>
         <shiro.version>1.4.0</shiro.version>
         <jackson.version>2.9.5</jackson.version>
